requirements.txt
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.28.0
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=9.5.0
